rename vector store package with lindorm and update related documents(modified files have been fully tested)
